# shop-api-golang/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateOrderItem:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  dto.CreateOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.CreateOrderItem'
        type: array
      user_id:
        example: 1
        type: integer
    type: object
  dto.CreateOrderResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/dto.OrderItemResponse'
        type: array
      user_id:
        type: integer
    required:
    - created_at
    - id
    - items
    - user_id
    type: object
...
  dto.DailyPurchases:
    properties:
      order_date:
        type: string
      purchases:
        type: integer
    required:
    - order_date
    - purchases
    type: object
  dto.DailyStatResponse:
    properties:
      date:
        example: "2026-05-11 01:45:24.864701"
        type: string
      orders:
        example: 1
        type: integer
      revenue:
        example: 120
        type: number
    required:
    - date
    - orders
    - revenue
    type: object
...
  dto.ListDailyPurchasesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.DailyPurchases'
        type: array
      total:
        type: integer
    required:
    - data
    - total
    type: object
  dto.ListDailyStatResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.DailyStatResponse'
        type: array
      total:
        type: integer
    required:
    - data
    - total
    type: object
  dto.ListOrderResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.OrderResponse'
        type: array
      total:
        type: integer
    required:
    - data
    - total
    type: object
...
  dto.OrderItemResponse:
    properties:
      price:
        type: number
      product_id:
        type: integer
      quantity:
        type: integer
      title:
        type: string
    required:
    - price
    - product_id
    - quantity
    - title
    type: object
  dto.OrderResponse:
    properties:
      created_at:
        example: "2026-05-11 01:45:24.864701"
        type: string
      id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    required:
    - created_at
    - id
    - user_id
    type: object
...
  dto.StatsOrderResponse:
    properties:
      averageCheck:
        example: 1.3
        type: number
      total:
        example: 1
        type: integer
      totalThisMonth:
        example: 4.2
        type: number
    required:
    - averageCheck
    - total
    - totalThisMonth
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Educational SQL workshop backend API
  title: Shop API
  version: "1.0"
paths:
...
  /orders:
    get:
      description: Returns all orders from the database
      parameters:
      - description: Number of products to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Number of products to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListOrderResponse'
      security:
      - BearerAuth: []
      summary: List all orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Creates an order and publishes order.created event to Kafka
      parameters:
      - description: Order data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateOrderResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - orders
  /orders/{orderID}/items:
    get:
      description: Returns all items for a given order
      parameters:
      - description: Order ID
        in: path
        name: orderID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.OrderItemResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get order items
      tags:
      - orders
...
  /orders/daily-purchases:
    get:
      description: Returns all orders from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListDailyPurchasesResponse'
      security:
      - BearerAuth: []
      summary: List all daily purchases
      tags:
      - orders
  /orders/daily-stats:
    get:
      description: Returns daily stats from the database
      parameters:
      - description: Number of products to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Number of products to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListDailyStatResponse'
      security:
      - BearerAuth: []
      summary: List daily stats
      tags:
      - orders
  /orders/stats:
    get:
      description: Returns statistics orders from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StatsOrderResponse'
      security:
      - BearerAuth: []
      summary: Statitstics all orders
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"